src/app/unit-test/unit-test.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';

import { UnitTestComponent } from './unit-test.component';

describe('UnitTestComponent', () => {
  let component: UnitTestComponent;
  let fixture: ComponentFixture<UnitTestComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [ UnitTestComponent ]
    })
    .compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(UnitTestComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should return 0 if input is negative', () => {

    const result = component.checkCompute(1);
    expect(result).toBe(2);

  });

  it('should return completed if name is Secupay', () => {

    const result = component.checkComparison('Secupay');
    expect(result).toContain('Welcome to Software Development!');

  });
  





})
